% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checking.plots.R
\name{checking.plots}
\alias{checking.plots}
\title{Checking Plots}
\usage{
checking.plots(model, n.id = 3, COL = c("#0080FF", "#A9E2FF"))
}
\arguments{
\item{model}{a linear model object}

\item{n.id}{the number of points to identify}

\item{COL}{vector of two colors}
}
\description{
Function that creates the three graphs discussed in Sections 11.5.1, 11.5.2,
and 11.5.3 that assesses independence, normality, and constant variance
}
\note{
Requires package \pkg{MASS}.
}
\examples{
# Figure 11.10
mod.aov <- aov(StopDist~tire, data = Tire)
checking.plots(mod.aov)
rm(mod.aov)
}
\author{
Alan T. Arnholt
}
\seealso{
\code{\link{twoway.plots}}
}
\keyword{regression}

