% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{CircuitDesigns}
\alias{CircuitDesigns}
\title{Circuit Design Lifetime}
\format{
A data frame with 26 observations on the following 2 variables:
\describe{ 
\item{lifetime}{lifetimes in thousands of hours}
\item{design}{a factor with levels \code{Design1}, \code{Design2},
\code{Design3}, and \code{Design4}} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
\code{CircuitDesigns} contains the results from an accelerated life test
used to estimate the lifetime of four different circuit designs (lifetimes
in thousands of hours).
}
\examples{

bwplot(design ~ lifetime, data = CircuitDesigns)

}
\keyword{datasets}
