% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Engineer}
\alias{Engineer}
\title{Engineer Salaries}
\format{
A data frame with 51 observations on the following 2 variables:
\describe{ 
\item{salary}{salary 10 years after graduation in thousands of dollars} 
\item{university}{one of three different engineering universities} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Salaries for engineering graduates 10 years after graduation
}
\examples{

boxplot(salary ~ university, data = Engineer, horizontal = TRUE)
# Trellis Approach
bwplot(university ~ salary, data = Engineer)

}
\keyword{datasets}
