% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{EURD}
\alias{EURD}
\title{European Union Research and Development}
\format{
A data frame with 15 observations on the following 3 variables:
\describe{ 
\item{Country}{a factor with levels \code{Bulgaria},
\code{Croatia}, \code{Cyprus}, \code{Czech Republic}, \code{Estonia},
\code{France}, \code{Hungary}, \code{Latvia}, \code{Lithuania},
\code{Malta}, \code{Portugal}, \code{Romania}, \code{Slovakia}, and
\code{Slovenia}} 
\item{RD2002}{research and development investments
in millions of euros for 2002} 
\item{RD2003}{research and development investments in millions of euros for 2003} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
A random sample of 15 countries' research and development investments for
the years 2002 and 2003 is taken and the results in millions of euros are
stored in \code{EURD}.
}
\examples{

qqmath(~(RD2003 - RD2002), data = EURD, type=c("p", "r"))

}
\keyword{datasets}
