% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Fertilize}
\alias{Fertilize}
\title{Cross and Auto Fertilization}
\format{
A data frame with 15 observations on the following 2 variables:
\describe{ 
\item{cross}{height of plant in inches}
\item{self}{height of plant in inches} 
}
}
\source{
Darwin, C. (1876) \emph{The Effect of Cross and Self-Fertilization
in the Vegetable Kingdom}
}
\description{
Plants' heights in inches obtained from two seeds, one obtained by cross
fertilization and the other by auto fertilization, in two opposite but
separate locations of a pot are recorded.
}
\examples{

with(data = Fertilize, 
t.test(cross, self))

}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\keyword{datasets}
