% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{food}
\alias{food}
\title{Carrot Shear}
\format{
A data frame with 16 observations on the following 2 variables:
\describe{ 
\item{shear}{carrot shear measured in kN}
\item{freezer}{a factor with levels \code{A}, \code{B}, \code{C},
and \code{D}} }
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Shear measured in kN on frozen carrots from four randomly selected freezers
}
\examples{

summary(aov(shear ~ freezer, data = food))

}
\keyword{datasets}
