% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Formula1}
\alias{Formula1}
\title{Pit Stop Times}
\format{
A data frame with 10 observations on the following 3 variables:
\describe{ 
\item{Race}{number corresponding to a race site}
\item{Team1}{pit stop times for team one} 
\item{Team2}{pit stop times for team two} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Pit stop times for two teams at 10 randomly selected Formula 1 races
}
\examples{

with(data = Formula1, 
boxplot(Team1, Team2))

}
\keyword{datasets}
