% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{glucose}
\alias{glucose}
\title{Blood Glucose Levels}
\format{
A data frame with 15 observations on the following 3 variables:
\describe{ 
\item{Patient}{patient number} 
\item{Old}{blood glucose level in mg/100 ml using old device} 
\item{New}{blood glucose level in mg/100 ml using new device} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Fifteen diabetic patients were randomly selected, and their blood glucose
levels were measured in mg/100 ml with two different devices.
}
\examples{

with(data = glucose, 
boxplot(Old, New))

}
\keyword{datasets}
