% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{InsurQuotes}
\alias{InsurQuotes}
\title{Insurance Quotes}
\format{
A data frame with 15 observations on the following 2 variables:
\describe{ 
\item{companyA}{quotes from company A in euros}
\item{companyB}{quotes from company B in euros} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Insurance quotes for two insurers of hazardous waste jobs
}
\examples{

with(data = InsurQuotes, 
t.test(companyA, companyB))

}
\keyword{datasets}
