% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{janka}
\alias{janka}
\title{Australian Eucalypt Hardwoods}
\format{
A data frame with 36 observations on the following 2 variables.
\describe{ 
\item{Density}{a measure of density of the timber}
\item{Hardness}{the Janka hardness of the timber} 
}
}
\source{
Williams, E.J. (1959) \emph{Regression Analysis}.  John Wiley &
Sons, New York.
}
\description{
The dataset consists of density and hardness measurements from 36 Australian
Eucalypt hardwoods.
}
\details{
Janka Hardness is an importance rating of Australian hardwood timbers. The
test measures the force required to imbed a steel ball into a piece of wood.
}
\examples{

with(data = janka, plot(Hardness ~ Density, col = "blue"))

}
\keyword{datasets}
