% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksdist.R
\name{ksdist}
\alias{ksdist}
\title{Simulated Distribution of Dn (Kolmogorov-Smirnov)}
\usage{
ksdist(n = 10, sims = 10000, alpha = 0.05)
}
\arguments{
\item{n}{sample size}

\item{sims}{number of simulations to perform}

\item{alpha}{desired alpha level}
}
\description{
Function to visualize the sampling distribution of Dn and to find simulated
critical values.
}
\examples{
ksdist(n = 10, sims = 10000, alpha = 0.05)
}
\author{
Alan T. Arnholt
}
\seealso{
\code{\link{ksLdist}}
}
\keyword{distribution}

