% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{MilkCarton}
\alias{MilkCarton}
\title{Milk Carton Drying Times}
\format{
A data frame with 40 observations on the following 2 variables:
\describe{ 
\item{Hgallon}{drying time in seconds for half gallon containers} 
\item{Wgallon}{drying time in seconds for whole gallon containers} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
A plastics manufacturer makes two sizes of milk containers: half gallon and
gallon sizes. The time required for each size to dry is recorded in seconds
in the data frame \code{MilkCarton}.
}
\examples{

with(data = MilkCarton, 
boxplot(Hgallon, Wgallon))

}
\keyword{datasets}
