% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{package}
\name{PASWR-package}
\alias{PASWR-package}
\alias{PASWR}
\title{Probability and Statistics with R}
\description{
Data and functions for the book \emph{Probability and Statistics with R}
}
\details{
\tabular{ll}{ Package: \tab PASWR\cr Type: \tab Package\cr Version: \tab
1.2\cr Date: \tab 2016-02-24\cr License: \tab GPL (>=2) \cr } Comprehensive
and engineering-oriented, \emph{Probability and Statistics with R} provides
a thorough treatment of probability and statistics, clear and accessible
real-world examples, and fully detailed proofs.  The text provides
step-by-step explanations for numerous examples in R and S-PLUS for nearly
every topic covered, including both traditional and nonparametric
techniques. With a wide range of graphs to illustrate complex material as
well as a solutions manual, the book also offers an accompanying website
that features supporting information, including datasets, functions, and
other downloadable material. It is ideal for undergraduate students and for
engineers and scientists who must perform statistical analyses.
}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\author{
Alan T. Arnholt

Maintainer: <arnholtat@appstate.edu>
}
\keyword{package}
