% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Rat}
\alias{Rat}
\title{Rat Survival Time}
\format{
A data frame with 20 observations on the following variable:
\describe{ 
\item{survival.time}{number of weeks survived} 
}
}
\source{
Lawless, J. (1982) \emph{Statistical Models and Methods for Lifetime
Data}. John Wiley, New York.
}
\description{
The survival time in weeks of 20 male rats exposed to high levels of
radiation.
}
\examples{

with(data = Rat, 
EDA(survival.time))

}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\keyword{datasets}
