% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Salinity}
\alias{Salinity}
\title{Surface Water Salinity}
\format{
A data frame with 48 observations on the following variable:
\describe{ 
\item{salinity}{surface-water salinity measurements} 
}
}
\source{
Davis, J. (1986) \emph{Statistics and Data Analysis in Geology}.
John Wiley, New York.
}
\description{
Surface-water salinity measurements were taken in a bottom-sampling project
in Whitewater Bay, Florida. These data are stored in the data frame
\code{Salinity}.
}
\examples{

with(data = Salinity, 
EDA(salinity))

}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\keyword{datasets}
