% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{SimDataST}
\alias{SimDataST}
\title{Simulated Data (Predictors)}
\format{
A data frame with 200 observations on the following 5 variables:
\describe{ 
\item{Y1}{a numeric vector} 
\item{Y2}{a numeric vector} 
\item{x1}{a numeric vector} 
\item{x2}{a numeric vector} 
\item{x3}{a numeric vector} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Simulated data for five variables.  Data are used with Example 12.21.
}
\examples{

xyplot(Y1 ~ x1, data = SimDataST, type=c("p", "smooth"))

}
\keyword{datasets}
