% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{StatTemps}
\alias{StatTemps}
\title{Student Temperature}
\format{
A data frame with 34 observations on the following 3 variables:
\describe{ 
\item{temperature}{temperature in farenheit}
\item{gender}{a factor with levels \code{Female} and \code{Male}}
\item{class}{a factor with levels \code{8 a.m.} and \code{9 a.m.}}
 }
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
In a study conducted at Appalachian State University, students used digital
oral thermometers to record their temperatures each day they came to class.
A randomly selected day of student temperatures is provided in
\code{StatTemps}. Information is also provided with regard to subject gender
and the hour of the day when the students' temperatures were measured.
}
\examples{

bwplot(gender ~ temperature, data = StatTemps)

}
\keyword{datasets}
