% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Stschool}
\alias{Stschool}
\title{School Satisfaction}
\format{
A data frame with 26 observations on the following 4 variables:
\describe{ 
\item{X}{satisfaction score}
\item{Y}{satisfaction score} 
\item{Satisfaction}{combined satisfaction scores} 
\item{School}{a factor with levels \code{X} and \code{Y}} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
A questionnaire is randomly administered to 11 students from State School
\code{X} and to 15 students from State School \code{Y} (the results have
been ordered and stored in the data frame \code{Stschool}).  Data are used
in Example 9.11.
}
\examples{

with(data = Stschool, 
t.test(X, Y, var.equal = TRUE))

}
\keyword{datasets}
