% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{TestScores}
\alias{TestScores}
\title{Statistics Grades}
\format{
A data frame with 29 observations on the following variable:
\describe{ 
\item{grade}{test score}
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Test grades of 29 students taking a basic statistics course
}
\examples{

with(data = TestScores,
EDA(grade))

}
\keyword{datasets}
