% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{TOE}
\alias{TOE}
\title{Nuclear Energy}
\format{
A data frame with 12 observations on the following variable:
\describe{ 
\item{energy}{nuclear energy measured in tons of oil equivalent} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Nuclear energy (in TOE, tons of oil equivalent) produced in 12 randomly
selected European countries during 2003
}
\examples{

with(TOE,
plot(density(energy)))

}
\keyword{datasets}
