% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{URLaddress}
\alias{URLaddress}
\title{Megabytes Downloaded}
\format{
A data frame with 30 observations on the following 2 variables:
\describe{ 
\item{megasd}{megabytes dowloaded}
\item{mconnected}{number of minutes connected} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
The manager of a URL commercial address is interested in predicting the
number of megabytes downloaded, \code{megasd}, by clients according to the
number minutes they are connected, \code{mconnected}. The manager randomly
selects (megabyte, minute) pairs, and records the data.  The pairs
(\code{megasd}, \code{mconnected}) are stored in the data frame
\code{URLaddress}.
}
\examples{

xyplot(mconnected ~ megasd, data = URLaddress, type=c("p", "r"))

}
\keyword{datasets}
