% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Wait}
\alias{Wait}
\title{Waiting Time}
\format{
A data frame with 15 observations on the following variable:
\describe{ 
\item{wt}{waiting time in minutes} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
A statistician records how long he must wait for his bus each morning.  Data
are used in Example 10.4.
}
\examples{

with(data = Wait,
wilcox.test(wt, mu = 6, alternative = "less"))

}
\keyword{datasets}
