% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Washer}
\alias{Washer}
\title{Washer Diameter}
\format{
A data frame with 20 observations on the following variable:
\describe{ 
\item{diameters}{diameter of washer in cm} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Diameter of washers.
}
\examples{

with(data = Washer,
EDA(diameters))

}
\keyword{datasets}
