\name{WCST}
\alias{WCST}
\docType{data}
\title{Wisconsin Card Sorting Test}
\description{
The following data are the test scores from a 
group of 50 patients from the
\emph{Virgen del Camino} Hospital (Pamplona, Spain) on the 
Wisconsin Card Sorting Test.
}
\usage{WCST}
\format{
  A data frame with 50 observations on the following variable:
  \describe{
    \item{\code{score}}{score on the Wisconsin Card Sorting Test}
  }
}
\details{
The \dQuote{Wisconsin Card Sorting Test} is widely used by psychiatrists, neurologists,
and neuropsychologists with patients who have a brain injury, neurodegenerative
disease, or a mental illness such as schizophrenia. Patients with any sort of
frontal lobe lesion generally do poorly on the test.
}

\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.}

\keyword{datasets}
