% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{AptSize}
\alias{AptSize}
\title{Apartment Size}
\format{
A data frame with 8 observations on the following 2 variables:
\describe{ \item{Mendebaldea}{Mendebaldea apartment size in square
meters} \item{SanJorge}{San Jorge apartment size in square meters} }
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Size of apartments in Mendebaldea, Spain and San Jorge, Spain
}
\examples{

with(data = AptSize,
boxplot(Mendebaldea, SanJorge) )

}
\keyword{datasets}
