% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Baberuth}
\alias{Baberuth}
\title{George Herman Ruth}
\format{
A data frame with 22 observations on the following 14 variables.
\describe{ 
\item{Year}{year in which the season occurred}
\item{Team}{team he played for \code{Bos-A}, \code{Bos-N}, or
\code{NY-A}} \item{G}{games played} \item{AB}{at bats}
\item{R}{runs scored} \item{H}{hits}
\item{X2B}{doubles} \item{X3B}{triples}
\item{HR}{home runs} \item{RBI}{runs batted in}
\item{SB}{stolen bases} \item{BB}{base on balls or walks}
\item{BA}{batting average H/AB} \item{SLG}{slugging percentage (total bases/at bats)} 
}
}
\source{
\url{https://www.baseball-reference.com/about/bat_glossary.shtml}
}
\description{
Baseball statistics for George Herman Ruth (The Bambino or The Sultan Of
Swat)
}
\examples{

with(data = Baberuth, 
hist(RBI))

}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\keyword{datasets}
