% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bino.gen.R
\name{bino.gen}
\alias{bino.gen}
\title{Simulating Binomial Distribution}
\usage{
bino.gen(samples, n, pi)
}
\arguments{
\item{samples}{number of repeated samples to generate}

\item{n}{number of Bernoulli trials}

\item{pi}{probability of success for Bernoulli trial}
}
\value{
\item{simulated.distribution}{Simulated binomial distribution}
\item{theoretical.distribution}{Theoretical binomial distribution}
}
\description{
Function that generates and displays \emph{m} repeated samples of \emph{n}
Bernoulli trials with a given probability of success.
}
\examples{

bino.gen(1000, 20, 0.75)

}
\author{
Alan T. Arnholt
}
\keyword{distribution}
