% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Calculus}
\alias{Calculus}
\title{Calculus Assessment Scores}
\format{
A data frame with 18 observations on the following 2 variables:
\describe{ 
\item{No.Calculus}{assessment score for students with no prior calculus} 
\item{Yes.Calculus}{assessment score for students with prior calculus} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Mathematical assessment scores for 36 students enrolled in a biostatistics
course according to whether or not the students had successfully completed a
calculus course prior to enrolling in the biostatistics course
}
\examples{

with(data = Calculus,
z.test(x = Yes.Calculus, y = No.Calculus, sigma.x = 5, sigma.y = 12)$conf
)

}
\keyword{datasets}
