% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.plots.R
\name{checking.plots}
\alias{checking.plots}
\title{Checking Plots}
\usage{
checking.plots(model, n.id = 3, COL = c("#0080FF", "#A9E2FF"))
}
\arguments{
\item{model}{an aov or lm object}

\item{n.id}{the number of points to identify}

\item{COL}{vector of two colors}
}
\description{
Function that creates four graphs that can be used to help assess independence, normality, and constant variance
}
\examples{
mod.aov <- aov(StopDist ~ tire, data = Tire)
checking.plots(mod.aov)
rm(mod.aov)


}
\seealso{
\code{\link{twoway.plots}}, \code{\link{oneway.plots}}
}
\author{
Alan T. Arnholt <arnholtat@appstate.edu>
}
\keyword{hplot}
