\name{CircuitDesigns}
\alias{CircuitDesigns}
\docType{data}
\title{Circuit Design Lifetime}
\description{
\code{CircuitDesigns} contains the results from an accelerated life 
test used to estimate the lifetime of four different circuit designs 
(lifetimes in thousands of hours).
}
\usage{CircuitDesigns}
\format{
  A data frame with 26 observations on the following 2 variables:
  \describe{
    \item{\code{lifetime}}{lifetimes in thousands of hours}
    \item{\code{design}}{a factor with levels \code{Design1}, 
    \code{Design2}, \code{Design3}, and \code{Design4}}
  }
}

\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.}

\keyword{datasets}
