% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Depend}
\alias{Depend}
\title{Number of Dependent Children for 50 Families}
\format{
A data frame with 50 observations on the following 4 variables.
\describe{ 
\item{C1}{a numeric vector} 
\item{number}{a numeric vector} 
\item{Count}{a numeric vector}
\item{Percent}{a numeric vector} }
}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Number of dependent children for 50 families.
}
\examples{

with(data = Depend, table(C1))

}
\keyword{datasets}
