% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Grocery}
\alias{Grocery}
\title{Grocery Spending}
\format{
A data frame with 30 observations on the following variable:
\describe{ 
\item{groceries}{total dollars spent on groceries} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
The consumer expenditure survey, created by the U.S. Department of Labor,
was administered to 30 households in Watauga County, North Carolina, to see
how the cost of living in Watauga county with respect to total dollars spent
on groceries compares with other counties. The amount of money each
household spent per week on groceries is stored in the variable
\code{groceries}.  Data are used in Example 8.3.
}
\examples{

with(data = Grocery, 
z.test(x = groceries, sigma.x = 30, conf.level = 0.97)$conf)

}
\keyword{datasets}
