% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{House}
\alias{House}
\title{House Prices}
\format{
A data frame with 14 observations on the following 2 variables:
\describe{ 
\item{Neighborhood}{a factor with levels \code{Blowing Rock}, \code{Cove Creek}, 
\code{Green Valley}, \code{Park Valley}, \code{Parkway}, and \code{Valley Crucis}} 
\item{Price}{price of house (in thousands of dollars)} 
}
}
\description{
Random sample of house prices (in thousands of dollars) for three
bedroom/two bath houses in Watauga County, NC
}
\examples{

with(data = House, 
t.test(Price))

}
\keyword{datasets}
