% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Kinder}
\alias{Kinder}
\title{Kindergarten Class}
\format{
A data frame with 20 observations on the following 2 variables:
\describe{ 
\item{ht}{height in inches of child}
\item{wt}{weight in pounds of child} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
The data frame \code{Kinder} contains the height in inches and weight in
pounds of 20 children from a kindergarten class.  Data are used in Example
12.17.
}
\examples{

# Figure 12.10
with(data = Kinder, 
plot(wt, ht))
# Trellis Approach
xyplot(ht ~ wt, data = Kinder)

}
\keyword{datasets}
