% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{LostR}
\alias{LostR}
\title{Lost Revenue Due to Worker Illness}
\format{
A data frame with 25 observations on the following 2 variables:
\describe{ 
\item{NumberSick}{number of absent workers due to illness} 
\item{LostRevenue}{lost revenue in dollars} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Data set containing the lost revenue in dollars/day and number of workers
absent due to illness for a metallurgic company
}
\examples{

xyplot(LostRevenue ~ NumberSick, data = LostR, type=c("p", "r"))

}
\keyword{datasets}
