% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.plots.R
\name{oneway.plots}
\alias{oneway.plots}
\title{Exploratory Graphs for Single Factor Designs}
\usage{
oneway.plots(Y, fac1, COL = c("#A9E2FF", "#0080FF"))
}
\arguments{
\item{Y}{response variable for a single factor design}

\item{fac1}{predictor variable (factor)}

\item{COL}{a vector with two colors}
}
\description{
Function to create dotplots, boxplots, and design plot (means) for single factor designs
}
\examples{
with(data = Tire, oneway.plots(StopDist, tire))

}
\seealso{
\code{\link{twoway.plots}}, \code{\link{checking.plots}}
}
\author{
Alan T. Arnholt <arnholtat@appstate.edu>
}
\keyword{hplot}
