% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Phone}
\alias{Phone}
\title{Telephone Call Times}
\format{
A data frame with 23 observations on the following variable:
\describe{ 
\item{call.time}{time spent on long distance calls in minutes} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
\code{Phone} contains times in minutes of long distance telephone calls
during a one month period for a small business.  Data are used in Example
10.1.
}
\examples{

with(data = Phone, 
SIGN.test(call.time, md = 2.1))

}
\keyword{datasets}
