\name{Schizo}
\alias{Schizo}
\docType{data}
\title{Dopamine Activity}

\description{
Twenty-five patients with schizophrenia were classified as psychotic
or nonpsychotic after being treated with an antipsychotic drug.  
Samples of cerebral fluid were taken from each patient and assayed
for dopamine b-hydroxylase (DBH) activity. The dopamine measurements
for the two groups are in nmol/(ml)(h)/(mg) of protein.
}
\usage{Schizo}
\format{
  A data frame with 15 observations on the following 2 variables:
  \describe{
    \item{\code{nonpsychotic}}{dopamine activity level for patients 
    classified nonpsychotic}
    \item{\code{psychotic}}{dopamine activity level for patients 
    classified psychotic}
  }
}

\source{
Sternberg, D. E., Van Kammen, D. P., and Bunney,W. E. (1982) 
\dQuote{Schizophrenia: Dopamine
b-Hydroxylase Activity and Treatment Response.} \emph{Science},
\bold{216}: 1423-1425.}

\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.}

\keyword{datasets}
