% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Score}
\alias{Score}
\title{Standardized Test Scores}
\format{
A data frame with 20 observations on the following variable:
\describe{ 
\item{scores}{standardized test score} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Standardized test scores from a random sample of twenty college freshmen.
}
\examples{

qqmath(~scores, data = Score, type=c("p", "r"))

}
\keyword{datasets}
