% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{SimDataXT}
\alias{SimDataXT}
\title{Simulated Data (Logarithms)}
\format{
A data frame with 200 observations on the following 4 variables:
\describe{ 
\item{Y}{a numeric vector} 
\item{x1}{a numeric vector} 
\item{x2}{a numeric vector} 
\item{x3}{a numeric vector} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Simulated data for four varaibles.  Data are used with Example 12.18.
}
\examples{

xyplot(Y ~ x1, data = SimDataXT, type=c("p", "smooth"))

}
\keyword{datasets}
