% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{TireWear}
\alias{TireWear}
\title{Tire Wear}
\format{
A data frame with 16 observations on the following 3 variables:
\describe{ 
\item{Wear}{tread loss measured in thousandths of an inch} 
\item{Treat}{a factor with levels \code{A}, \code{B},
\code{C}, and \code{D}} 
\item{Block}{a factor with levels \code{Car1}, \code{Car2}, \code{Car3}, and \code{Car4}} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
The data frame \code{TireWear} contains measurements for the amount of tread
loss after 10,000 miles of driving in thousandths of an inch. Data are used
in Example 11.8.
}
\examples{

par(mfrow = c(1, 2), cex = 0.8)
with(data = TireWear,
interaction.plot(Treat, Block, Wear, type = "b", legend = FALSE))
with(data = TireWear, 
interaction.plot(Block, Treat, Wear, type = "b", legend = FALSE))
par(mfrow = c(1, 1), cex = 1)

}
\keyword{datasets}
