\name{TOE}
\alias{TOE}
\docType{data}
\title{Nuclear Energy}
\description{
Nuclear energy (in TOE, tons of oil equivalent)
produced in 12 randomly selected European countries during 2003
}
\usage{TOE}
\format{
  A data frame with 12 observations on the following variable:
  \describe{
    \item{\code{energy}}{nuclear energy measured in tons of oil equivalent}
  }
}

\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.}


\keyword{datasets}
