% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Top20}
\alias{Top20}
\title{Tennis Income}
\format{
A data frame with 15 observations on the following variable:
\describe{ 
\item{income}{yearly income in millions of dollars} 
}
}
\source{
\url{https://www.atptour.com/}
}
\description{
\code{Top20} contains data (in millions of dollars) corresponding to the
earnings of 15 randomly selected tennis players whose earnings fall
somewhere in positions 20 through 100 of ranked earnings.
}
\examples{

with(data = Top20, 
EDA(income))

}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\keyword{datasets}
