% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Water}
\alias{Water}
\title{Sodium Content of Water}
\format{
A data frame with 30 observations on the following 4 variables:
\describe{ 
\item{X}{sodium content measured in mg/L}
\item{Y}{sodium content measured in mg/L}
\item{Sodium}{combined sodium content measured in mg/L}
\item{Source}{a factor with levels \code{X} and \code{Y}} }
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
An independent agency measures the sodium content in 20 samples from source
\code{X} and in 10 samples from source \code{Y} and stores them in data
frame \code{Water}. Data are used in Example 9.12.
}
\examples{

with(data = Water,
t.test(X, Y, alternative = "less"))

}
\keyword{datasets}
