% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{WCST}
\alias{WCST}
\title{Wisconsin Card Sorting Test}
\format{
A data frame with 50 observations on the following variable:
\describe{ 
\item{score}{score on the Wisconsin Card Sorting Test} 
}
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
The following data are the test scores from a group of 50 patients from the
\emph{Virgen del Camino} Hospital (Pamplona, Spain) on the Wisconsin Card
Sorting Test.
}
\details{
The \dQuote{Wisconsin Card Sorting Test} is widely used by psychiatrists,
neurologists, and neuropsychologists with patients who have a brain injury,
neurodegenerative disease, or a mental illness such as schizophrenia.
Patients with any sort of frontal lobe lesion generally do poorly on the
test.
}
\examples{

densityplot(~score, data = WCST, ref = TRUE)

}
\keyword{datasets}
