% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASWR-package.R
\docType{data}
\name{Wool}
\alias{Wool}
\title{Wool Production}
\format{
A data frame with 15 observations on the following 2 variables:
\describe{ 
\item{textileA}{wool production in thousands of kilograms}
 \item{textileB}{wool production in thousands of kilograms} 
 }
}
\source{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008)
\emph{Probability and Statistics with R}. Chapman & Hall/CRC.
}
\description{
Random sample of wool production in kilograms on 5 different days at two
different locations
}
\examples{

with(data = Wool, 
t.test(textileA, textileB))

}
\keyword{datasets}
